+++ docs/src/app/(website)/layout.tsx
@@ -7,7 +7,7 @@ import type { Metadata, Viewport } from 'next/types';
 import { GoogleAnalytics } from 'docs/src/components/GoogleAnalytics';
 import { Link } from 'docs/src/components/Link';
 import { Logo } from 'docs/src/components/Logo';
-import { SearchTrigger } from 'docs/src/components/SearchTrigger';
+import { SearchControls } from 'docs/src/components/SearchControls';
 
 export default function Layout({ children }: React.PropsWithChildren) {
   return (
@@ -54,7 +54,7 @@ export default function Layout({ children }: React.PropsWithChildren) {
                 </Link>
               </nav>
               <div className="bui-d-n bp3:bui-d-f bui-fd-c bui-g-2 bui-ai-s bui-gcs-7 bui-gce-9">
-                <SearchTrigger enableKeyboardShortcut />
+                <SearchControls />
               </div>
             </header>
             <main id="main" className="bui-d-c">
