--- docs/src/components/Search/index.ts
@@ -1,1 +1,0 @@
-export * from './Search';
